pyproject.toml
[project]
name = "lynne"
version = "0.1.0"
description = "个人社交平台信息自动化采集与分析工具"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "loguru>=0.7",
    "playwright>=1.48",
    "playwright-stealth>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "apscheduler>=3.10",
    "jinja2>=3.1",
    "openai>=1.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.34"]

[project.scripts]
lynne = "src.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
